Refactored database tests
