+++ src/test/java/DataBaseTest.java
@@ -1,10 +1,7 @@
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-import org.json.simple.parser.ParseException;
+import DataBase.DataBase;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import Entities.*;
-import DataBase.*;
 import java.util.ArrayList;
 import java.util.Date;
 public class DataBaseTest {
@@ -13,8 +10,7 @@ public class DataBaseTest {
     String testDate = "Fri Nov 18 01:04:05 EST 2022";
     Item plushie = new Item("Plushie", 40.99, 30.00, "www.amazon.com/plushie",
             "Description from amazon (or you write your own)", new String[]{"toys"}, 40.99, new Date(testDate), 0, 0, "www.amazonimage.com/keyboard", "CAD", priceData, priceDate);
-    ArrayList<Product> items = new ArrayList<>();
-    ArrayList<String> tags = new ArrayList<>();
+
     @Test
     public void TestDataBaseReturnsUserFilePath() {
         Assertions.assertEquals(DataBase.getUserFilePath(), "src/main/database/users.txt");
@@ -24,81 +20,7 @@ public class DataBaseTest {
         Assertions.assertEquals(DataBase.getWishlistPath("Herman1"), "src/main/database/Herman1.txt");
     }
     @Test
-    public void TestDataBaseReturnsItemJSON() {
-        DataBaseFormatter dataBaseFormatter = new DataBaseFormatter();
-        plushie.setDateAdded(new Date(testDate));
-        String itemString = dataBaseFormatter.createItemJSON(plushie).toString();
-        Assertions.assertTrue(itemString.contains("\"priceChange\":40.99"));
-        Assertions.assertTrue(itemString.contains("\"itemName\":\"Plushie\""));
-        Assertions.assertTrue(itemString.contains("\"desiredPrice\":30.0"));
-        Assertions.assertTrue(itemString.contains("\"reviewCount\":0"));
-        Assertions.assertTrue(itemString.contains("\"imageURL\":\"www.amazonimage.com\\/keyboard\""));
-        Assertions.assertTrue(itemString.contains("\"itemPrice\":40.99"));
-        Assertions.assertTrue(itemString.contains("\"itemDescription\":\"Description from amazon (or you write your own)\""));
-        Assertions.assertTrue(itemString.contains("\"reviewStars\":0.0,"));
-        Assertions.assertTrue(itemString.contains("\"url\":\"www.amazon.com\\/plushie\""));
+    public void TestDataBaseReturnsTempUserPath() {
+        Assertions.assertEquals(DataBase.getUserFilePath(), "src/main/database/tempusers.txt");
     }
-    @Test
-    public void TestDataBaseParsesItemObject() throws ParseException, java.text.ParseException {
-        DataBaseParser dataBaseParser = new DataBaseParser();
-        plushie.setDateAdded(new Date(testDate));
-        JSONParser jsonParser = new JSONParser();
-        JSONObject itemObject = (JSONObject) jsonParser.parse("{\"priceChange\":40.99,\"itemName\":\"Plushie\",\"desiredPrice\":30.0,\"reviewCount\":0,\"imageURL\":\"www.amazonimage.com\\/keyboard\",\"itemPrice\":40.99,\"itemDescription\":\"Description from amazon (or you write your own)\",\"reviewStars\":0.0,\"url\":\"www.amazon.com\\/plushie\",\"dateAdded\":\"Fri Nov 18 01:04:05 EST 2022\",\"tags\":[\"toys\"]}");
-        Product parsedItem = dataBaseParser.parseItem(itemObject);
-        Assertions.assertEquals(parsedItem.getProductDescription(), plushie.getProductDescription());
-        Assertions.assertEquals(parsedItem.getProductPrice(), plushie.getProductPrice());
-        Assertions.assertEquals(parsedItem.getProductName(), plushie.getProductName());
-        Assertions.assertEquals(parsedItem.getProductURL(), plushie.getProductURL());
-        Assertions.assertEquals(parsedItem.getProductDesiredPrice(), plushie.getProductDesiredPrice());
-        Assertions.assertArrayEquals(parsedItem.getTags(), plushie.getTags());
-        Assertions.assertEquals(parsedItem.getPriceChange(), plushie.getPriceChange());
-        Assertions.assertEquals(parsedItem.getProductImageURL(), plushie.getProductImageURL());
-        Assertions.assertEquals(parsedItem.getReviewCount(), plushie.getReviewCount());
-        Assertions.assertEquals(parsedItem.getReviewStars(), plushie.getReviewStars());
-        Assertions.assertEquals(parsedItem.getProductDateAdded(), plushie.getProductDateAdded());
-    }
-    @Test
-    public void TestDataBaseReturnsWishListJSON() {
-        DataBaseFormatter dataBaseFormatter = new DataBaseFormatter();
-        plushie.setDateAdded(new Date(testDate));
-        items.add(plushie);
-        tags.add("Blue");
-        Wishlist wishlist = new Wishlist("Singles Day List", items, items, new Date(testDate), tags);
-        String wishlistString = dataBaseFormatter.createWishlistJSON(wishlist).toString();
-        Assertions.assertTrue(wishlistString.contains("\"name\":\"Singles Day List\""));
-        Assertions.assertTrue(wishlistString.contains("\"displayedList\""));
-        Assertions.assertTrue(wishlistString.contains("\"itemList\""));
-        Assertions.assertTrue(wishlistString.contains("\"selectedTags\""));
-    }
-    @Test
-    public void TestDataBaseParsesWishListObject() throws ParseException, java.text.ParseException {
-        DataBaseParser dataBaseParser = new DataBaseParser();
-        plushie.setDateAdded(new Date(testDate));
-        items.add(plushie);
-        tags.add("Blue");
-        Wishlist wishlist = new Wishlist("Singles Day List", items, items, new Date(testDate), tags);
-        JSONParser jsonParser = new JSONParser();
-        JSONObject listObject = (JSONObject) jsonParser.parse("{\"displayedList\":[{\"priceChange\":40.99,\"itemName\":\"Plushie\",\"desiredPrice\":30.0,\"reviewCount\":0,\"imageURL\":\"www.amazonimage.com\\/keyboard\",\"itemPrice\":40.99,\"itemDescription\":\"Description from amazon (or you write your own)\",\"reviewStars\":0.0,\"url\":\"www.amazon.com\\/plushie\",\"dateAdded\":\"Fri Nov 18 01:04:05 EST 2022\",\"tags\":[\"toys\"]}],\"name\":\"Singles Day List\",\"selectedTags\":[\"Blue\"],\"itemList\":[{\"priceChange\":40.99,\"itemName\":\"Plushie\",\"desiredPrice\":30.0,\"reviewCount\":0,\"imageURL\":\"www.amazonimage.com\\/keyboard\",\"itemPrice\":40.99,\"itemDescription\":\"Description from amazon (or you write your own)\",\"reviewStars\":0.0,\"url\":\"www.amazon.com\\/plushie\",\"dateAdded\":\"Fri Nov 18 01:04:05 EST 2022\",\"tags\":[\"toys\"]}],\"dateAdded\":\"Fri Nov 18 01:04:05 EST 2022\"}");
-        Wishlist parsedList = dataBaseParser.parseWishlist(listObject);
-        Assertions.assertEquals(parsedList.getName(), wishlist.getName());
-        Assertions.assertEquals(parsedList.getDateAdded(), wishlist.getDateAdded());
-        Assertions.assertEquals(parsedList.getDisplayedList().toArray().length, wishlist.getDisplayedList().toArray().length);
-        Assertions.assertArrayEquals(parsedList.getSelectedTags().toArray(), wishlist.getSelectedTags().toArray());
-        Assertions.assertEquals(parsedList.getItemList().toArray().length, wishlist.getItemList().toArray().length);
-    }
-
-    @Test
-    public void TestDataBaseWritesAndSavesListOfWishLists() {
-        DataBaseController dataBaseController = new DataBaseController();
-        ListOfWishlists wishlists = new ListOfWishlists();
-        Wishlist wishlist = new Wishlist("Exam Celebration Wish List");
-        wishlist.setDateAdded(new Date(testDate));
-        wishlists.addWishlist(wishlist);
-
-        User user = new User("TestUser", "Test");
-        DataBase.saveListOfWishlists(wishlists, user);
-
-        Assertions.assertEquals(dataBaseController.getListOfWishlists("TestUser").getListOfWishlist().get(0).getName(), "Exam Celebration Wish List");
-    }
-
 }
